[AJ-1296] Organize data tables UI part 2 (#4163)
