+++ src/components/data/EntitiesContent.js
@@ -10,9 +10,8 @@ import { tools } from 'src/analysis/utils/tool-utils';
 import { requesterPaysWrapper, withRequesterPaysHandler } from 'src/components/bucket-utils';
 import { ButtonSecondary } from 'src/components/common';
 import { DataTableColumnProvenance } from 'src/components/data/data-table-provenance';
-import { AddColumnModal, AddEntityModal, CreateEntitySetModal, EntityDeleter, ModalToolButton } from 'src/components/data/data-utils';
+import { ModalToolButton } from 'src/components/data/data-utils';
 import DataTable from 'src/components/data/DataTable';
-import ExportDataModal from 'src/components/data/ExportDataModal';
 import { icon, spinner } from 'src/components/icons';
 import IGVBrowser from 'src/components/IGVBrowser';
 import IGVFileSelector from 'src/components/IGVFileSelector';
@@ -22,7 +21,12 @@ import { withModalDrawer } from 'src/components/ModalDrawer';
 import { MenuDivider, MenuTrigger } from 'src/components/PopupTrigger';
 import TitleBar from 'src/components/TitleBar';
 import WorkflowSelector from 'src/components/WorkflowSelector';
+import { AddColumnModal } from 'src/data/data-table/entity-service/AddColumnModal';
+import { AddEntityModal } from 'src/data/data-table/entity-service/AddEntityModal';
+import { CreateEntitySetModal } from 'src/data/data-table/entity-service/CreateEntitySetModal';
 import { entityAttributeText } from 'src/data/data-table/entity-service/entityAttributeText';
+import { EntityDeleter } from 'src/data/data-table/entity-service/EntityDeleter';
+import { ExportDataModal } from 'src/data/data-table/entity-service/ExportDataModal';
 import { MultipleEntityEditor } from 'src/data/data-table/entity-service/MultipleEntityEditor';
 import datasets from 'src/data/datasets';
 import dataExplorerLogo from 'src/images/data-explorer-logo.svg';
